Refactor pages to separate folder
