+++ coffee-chats/src/main/webapp/src/App.tsx
@@ -1,11 +1,11 @@
 import React from "react";
 import "./App.css";
-import {MainPage} from "./components/MainPage";
+import {MainPage} from "./pages/MainPage";
 import {NavBar} from "./components/NavBar";
 import { MuiPickersUtilsProvider } from "@material-ui/pickers"
 import DateFnsUtils from "@date-io/date-fns"
 import {HashRouter as Router, Switch, Route} from "react-router-dom";
-import {GroupListPage} from "./components/GroupListPage";
+import {GroupListPage} from "./pages/GroupListPage";
 
 function App() {
   return (
