+++ coffee-chats/src/main/webapp/src/pages/GroupListPage.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import {Box, Container, Grid, TextField} from "@material-ui/core";
-import {GroupCard} from "./GroupCard";
+import {GroupCard} from "../components/GroupCard";
 
 export function GroupListPage() {
   return (
